// generated/models/OneDriveforBusinessModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface DataSetsMetadata {
  tabular?: TabularDataSetsMetadata;
  blob?: BlobDataSetsMetadata;
}

export interface TabularDataSetsMetadata {
  /** Dataset source */
  source?: string;
  /** Dataset display name */
  displayName?: string;
  /** Dataset url encoding */
  urlEncoding?: string;
  /** Table display name */
  tableDisplayName?: string;
  /** Table plural display name */
  tablePluralName?: string;
}

export interface BlobDataSetsMetadata {
  /** Blob dataset source */
  source?: string;
  /** Blob dataset display name */
  displayName?: string;
  /** Blob dataset url encoding */
  urlEncoding?: string;
}

export interface BlobMetadata {
  /** The unique identifier of the file or folder. */
  Id?: string;
  /** The name of the file or folder. */
  Name?: string;
  /** The name of the file or folder, without the file extension. */
  NameNoExt?: string;
  /** The display name of the file or folder. */
  DisplayName?: string;
  /** The path of the file or folder. */
  Path?: string;
  /** The date and time the file or folder was last modified. */
  LastModified?: string;
  /** The file or folder size in bytes. */
  Size?: number;
  /** The media type of the file or folder. */
  MediaType?: string;
  /** A boolean value (true, false) to indicate whether or not the blob is a folder. */
  IsFolder?: boolean;
  /** The etag of the file or folder. */
  ETag?: string;
  /** The file locator of the file or folder. */
  FileLocator?: string;
  /** The user who last modified the file or folder. */
  LastModifiedBy?: string;
}

export interface SharingLink {
  /** A URL that points to the file or folder. */
  WebUrl?: string;
}

export interface Tags {
  /** A list of tags. */
  Tags?: string[];
}

export interface Thumbnail {
  /** A URL that points to the thumbnail. */
  Url?: string;
  /** The thumbnail width in pixels. */
  Width?: number;
  /** The thumbnail height in pixels. */
  Height?: number;
}

export interface ForASelectedFileResponse {
  /** The unique path of the file. */
  filePath?: string;
  /** The URL to the source file. */
  fileUrl?: string;
}

export interface BlobMetadataPage {
  /** Blob metadata collection. */
  value?: BlobMetadata[];
  /** A URL which can be used to retrieve the next page. */
  nextLink?: string;
}